+++ app/Console/Commands/PromoteAdmin.php
@@ -3,7 +3,7 @@
 namespace App\Console\Commands;
 
 use App\Role;
-use App\User;
+use App\Models\User;
 use Illuminate\Console\Command;
 
 class PromoteAdmin extends Command
