+++ app/Models/Answer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,6 +8,6 @@ class Answer extends Model
 {
     public function question()
     {
-        return $this->belongsTo(\App\Question::class);
+        return $this->belongsTo(\App\Models\Question::class);
     }
 }
