+++ app/Models/Set.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,11 +8,11 @@ class Set extends Model
 {
     public function questions()
     {
-        return $this->hasMany(\App\Question::class);
+        return $this->hasMany(\App\Models\Question::class);
     }
 
     public function tests()
     {
-        return $this->hasMany(\App\Test::class);
+        return $this->hasMany(\App\Models\Test::class);
     }
 }
