+++ app/Models/Test.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
@@ -9,17 +9,17 @@ class Test extends Model
 {
     public function questions()
     {
-        return $this->belongsToMany(\App\Question::class, 'test_question');
+        return $this->belongsToMany(\App\Models\Question::class, 'test_question');
     }
 
     public function set()
     {
-        return $this->belongsTo(\App\Set::class);
+        return $this->belongsTo(\App\Models\Set::class);
     }
 
     public function user()
     {
-        return $this->belongsTo(\App\User::class);
+        return $this->belongsTo(\App\Models\User::class);
     }
 
     public function getEndAtAttribute($time)
